working diagnostic test app, blocking winmd generator via return as getdiagnostics dont show mine
